[review] 헥사고날에서 포트 인터페이스 모듈 위치에 대한 제안
