--- api-in/src/main/java/me/gogosing/board/application/port/in/UpdateBoardArticleUseCase.java
@@ -1,14 +1,0 @@
-package me.gogosing.board.application.port.in;
-
-import javax.validation.Valid;
-import javax.validation.constraints.Min;
-import me.gogosing.board.application.port.in.request.command.UpdateBoardArticleInCommand;
-import me.gogosing.board.application.port.in.response.UpdateBoardArticleInResponse;
-
-public interface UpdateBoardArticleUseCase {
-
-	UpdateBoardArticleInResponse updateBoardArticle(
-		final @Min(1L) Long id,
-		final @Valid UpdateBoardArticleInCommand inCommand
-	);
-}
