+++ application/src/main/java/me/gogosing/board/application/CreateBoardArticleService.java
@@ -4,13 +4,13 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import me.gogosing.board.application.port.in.CreateBoardArticleUseCase;
+import me.gogosing.board.domain.ports.inside.CreateBoardArticleUseCase;
 import me.gogosing.board.application.port.in.request.command.CreateBoardArticleInCommand;
 import me.gogosing.board.application.port.in.request.command.CreateBoardAttachmentInCommand;
 import me.gogosing.board.application.port.in.response.CreateBoardArticleInResponse;
 import me.gogosing.board.application.port.in.response.CreateBoardAttachmentInResponse;
-import me.gogosing.board.application.port.out.CreateBoardArticlePort;
-import me.gogosing.board.application.port.out.CreateBoardAttachmentsPort;
+import me.gogosing.board.domain.ports.outside.CreateBoardArticlePort;
+import me.gogosing.board.domain.ports.outside.CreateBoardAttachmentsPort;
 import me.gogosing.board.domain.BoardAttachmentDomainEntity;
 import me.gogosing.board.domain.BoardDomainEntity;
 import me.gogosing.support.jta.JtaTransactional;
