+++ application/src/main/java/me/gogosing/board/application/DeleteBoardArticleService.java
@@ -1,9 +1,9 @@
 package me.gogosing.board.application;
 
 import lombok.RequiredArgsConstructor;
-import me.gogosing.board.application.port.in.DeleteBoardArticleUseCase;
-import me.gogosing.board.application.port.out.DeleteBoardArticlePort;
-import me.gogosing.board.application.port.out.DeleteBoardAttachmentsPort;
+import me.gogosing.board.domain.ports.inside.DeleteBoardArticleUseCase;
+import me.gogosing.board.domain.ports.outside.DeleteBoardArticlePort;
+import me.gogosing.board.domain.ports.outside.DeleteBoardAttachmentsPort;
 import me.gogosing.support.jta.JtaTransactional;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
