+++ application/src/main/java/me/gogosing/board/application/GetBoardArticleService.java
@@ -3,11 +3,11 @@ package me.gogosing.board.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import me.gogosing.board.application.port.in.GetBoardArticleQuery;
+import me.gogosing.board.domain.ports.inside.GetBoardArticleQuery;
 import me.gogosing.board.application.port.in.response.GetBoardArticleInResponse;
 import me.gogosing.board.application.port.in.response.GetBoardAttachmentInResponse;
-import me.gogosing.board.application.port.out.LoadBoardArticlePort;
-import me.gogosing.board.application.port.out.LoadBoardAttachmentsPort;
+import me.gogosing.board.domain.ports.outside.LoadBoardArticlePort;
+import me.gogosing.board.domain.ports.outside.LoadBoardAttachmentsPort;
 import me.gogosing.board.domain.BoardAttachmentDomainEntity;
 import me.gogosing.board.domain.BoardDomainEntity;
 import me.gogosing.support.jta.JtaTransactional;
