+++ application/src/main/java/me/gogosing/board/application/UpdateBoardArticleService.java
@@ -5,15 +5,15 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import me.gogosing.board.application.port.in.UpdateBoardArticleUseCase;
+import me.gogosing.board.domain.ports.inside.UpdateBoardArticleUseCase;
 import me.gogosing.board.application.port.in.request.command.UpdateBoardArticleInCommand;
 import me.gogosing.board.application.port.in.request.command.UpdateBoardAttachmentInCommand;
 import me.gogosing.board.application.port.in.response.UpdateBoardArticleInResponse;
 import me.gogosing.board.application.port.in.response.UpdateBoardAttachmentInResponse;
-import me.gogosing.board.application.port.out.CreateBoardAttachmentsPort;
-import me.gogosing.board.application.port.out.DeleteBoardAttachmentsPort;
-import me.gogosing.board.application.port.out.LoadBoardArticlePort;
-import me.gogosing.board.application.port.out.UpdateBoardArticlePort;
+import me.gogosing.board.domain.ports.outside.CreateBoardAttachmentsPort;
+import me.gogosing.board.domain.ports.outside.DeleteBoardAttachmentsPort;
+import me.gogosing.board.domain.ports.outside.LoadBoardArticlePort;
+import me.gogosing.board.domain.ports.outside.UpdateBoardArticlePort;
 import me.gogosing.board.domain.BoardAttachmentDomainEntity;
 import me.gogosing.board.domain.BoardDomainEntity;
 import me.gogosing.support.jta.JtaTransactional;
