+++ domain/src/main/java/me/gogosing/board/domain/ports/inside/CreateBoardArticleUseCase.java
@@ -1,4 +1,4 @@
-package me.gogosing.board.application.port.in;
+package me.gogosing.board.domain.ports.inside;
 
 import javax.validation.Valid;
 import me.gogosing.board.application.port.in.request.command.CreateBoardArticleInCommand;
