+++ domain/src/main/java/me/gogosing/board/domain/ports/inside/GetPaginatedBoardArticleQuery.java
@@ -1,4 +1,4 @@
-package me.gogosing.board.application.port.in;
+package me.gogosing.board.domain.ports.inside;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
